* Fixed compiler warnings about const mismatch * Fixed compiler warnings about missing cast * Fixed compiler warnings about unused variables * Fixed compiler warnings about possible uninitialized variables * Fixed compiler warnings about printf format * Fixed compiler warnings about unreachable code

+++ libsrc/Wi/CLIsql2.c
@@ -1370,7 +1370,7 @@ virtodbc__SQLSetConnectOption (
 	    if (unbox (tpd->dtrx_info) == unbox (info->d_trx_id))
 	      {
 		trx_i = info;
-		dbg_printf (("found transaction %ld\n", dk_set_length (trx_i->d_trx_hdbcs)));
+		dbg_printf (("found transaction %ld\n", (long) dk_set_length (trx_i->d_trx_hdbcs)));
 		break;
 	      }
 	  }
